fix: delete session while refresh token

+++ src/features/user-platform/application/usecases/delete-session-except.usecase.ts
@@ -1,12 +1,17 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
-import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { UserDeviceSessionsRepository } from '../../repositories/user-device-sessions.repository';
 
 export class DeleteSessionExceptCommand {
   constructor(
     public deviceId: string,
     public userId: string,
+    public iat: number,
   ) {}
 }
 
@@ -17,7 +22,17 @@ export class DeleteSessionExceptUseCase
   constructor(private sessionsRepository: UserDeviceSessionsRepository) {}
 
   async execute(command: DeleteSessionExceptCommand): Promise<void> {
-    const { deviceId, userId } = command;
+    const { deviceId, userId, iat } = command;
+    const session = await this.sessionsRepository.findByDeviceId(deviceId);
+
+    if (!session) {
+      throw new UnauthorizedException();
+    }
+
+    if (session.issuedAt !== iat) {
+      throw new UnauthorizedException();
+    }
+
     await this.sessionsRepository.deleteAllSessionsExcept(deviceId, userId);
   }
 }
