+++ src/features/user-platform/application/usecases/delete-session.usecase.ts
@@ -1,12 +1,17 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
-import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { UserDeviceSessionsRepository } from '../../repositories/user-device-sessions.repository';
 
 export class DeleteSessionCommand {
   constructor(
     public deviceId: string,
     public userId: string,
+    public iat: number,
   ) {}
 }
 
@@ -29,6 +34,10 @@ export class DeleteSessionUseCase
 
     if (session.userId !== command.userId) {
       throw new ForbiddenException();
+    }
+
+    if (session.issuedAt !== command.iat) {
+      throw new UnauthorizedException();
     }
 
     await this.sessionsRepository.deleteSession(session.id);
